Cleaned up grass visuals
Following no longer sets nocollide, and instead stops characters from
colliding with each other if one's following the other.
